mis en place du code d'authentification des utilisateurs et mis en place de la structure de la base de données

+++ suaps-backend/database/migrations/2026_01_17_190847_create_evaluations_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         Schema::create('evaluations', function (Blueprint $table) {
             $table->id();
 
-            $table->float('note')->nullable();
+            $table->decimal('note', 4, 2)->nullable();
             
             $table->foreignId('moniteur_id')->constrained('moniteurs');
             $table->foreignId('etudiant_id')->constrained();
